Move random pose sampler to derived stage GenerateRandomPose
